Fixed an error in the noun turk_parse which truncated some sentences
